+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -37,6 +37,7 @@
 #include "Rule18p8Check.h"
 #include "Rule19p2Check.h"
 #include "Rule8p10Check.h"
+#include "Rule8p14Check.h"
 #include "Rule9p3Check.h"
 
 namespace clang::tidy {
@@ -103,6 +104,8 @@ public:
         "misrac-rule-19p2");
     CheckFactories.registerCheck<Rule8p10Check>(
         "misrac-rule-8p10");
+    CheckFactories.registerCheck<Rule8p14Check>(
+        "misrac-rule-8p14");
     CheckFactories.registerCheck<Rule9p3Check>(
         "misrac-rule-9p3");
   }