+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -383,6 +383,12 @@ New checks
   Detects issues in accordance with MISRA C rule 8.10:
   "`An inline function shall be declared with the static storage class`".
 
+- New :doc:`misrac-rule-8p14
+  <clang-tidy/checks/misrac/rule-8p14>` check.
+
+  Detects issues in accordance with MISRA C rule 8.14:
+  "`The restrict type qualifier shall not be used`".
+
 - New :doc:`misrac-rule-9p3
   <clang-tidy/checks/misrac/rule-9p3>` check.